+++ src/arguments/index.ts
@@ -1,3 +1,4 @@
+import { dofusCoordinates } from './dofus/coordinates';
 import { dofusDimension } from './dofus/dimension';
 import { dofusJob } from './dofus/job';
 import { dofusServer } from './dofus/server';
@@ -5,6 +6,7 @@ import { duration } from './time/duration';
 import { line } from './text/line';
 
 export const argumentTypes = {
+  dofusCoordinates,
   dofusDimension,
   dofusJob,
   dofusServer,