+++ tests/suites/arguments/index.ts
@@ -1,3 +1,4 @@
+export * from './dofus-coordinates';
 export * from './dofus-dimension';
 export * from './dofus-job';
 export * from './dofus-server';